refactor training release pipeline
